// src/types.d.ts
export type ColorSet = {
  50: string
  100: string
  200: string
  300: string
  400: string
  500: string
  600: string
  700: string
  800: string
  900: string
}

export type ColorNames = {
  lightBlue: string
  warmGray: string
  trueGray: string
  coolGray: string
  blueGray: string
}

export type ColorState = {
  primary: string
  secondary: string
  tertiary: string
  neutral: string
}